Dept cleaning
Changed X/Y coordinances in Greeting so that the view cart, dept and checkout buttons would be more visable.
 x from 50 to 60
 y from 80 to 40

Added the displayDepts();  in Greeting to return an new membership window signup after the save button has been selected.

Added the java.io.File class and code to the pantry through tupperware depts to adjust the .png files and directories.