Dept CLosed
WOrked on if statement for Dept closures due to lack of products. Not sure how to make the VBox connect with the Scene in an if statement.

Also Figure out overlay of pictures for Health and Beauty Dept. Still working on Image alignment.

Lastly changed some of the lines in the  Receipt display to do a bit of a different look.